[DFI-540] storage key types moved to a separate file
